# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto API
  description: API for WasaPhoto
  version: 1.0.0
...
components:
  schemas:
...
    UserUpdate:
      description: 'Values for updating an user, the user can choose to change there username,firstname,lastname and email'
      type: object
      properties:
        username:
          description: "username of the person"
          type: string
          pattern: "^.*?$"
          minLength: 3
          maxLength: 15
        firstname:
          description: "firstname of the person"
          type: string
          pattern:  "^.*?$"
          minLength: 3
          maxLength: 15
        lastname:
          description: "lastname of the person"
          type: string
          pattern: "^.*?$"
          minLength: 3
          maxLength: 15
        email:
          description: "email of the person"
          type: string
          format: email
          minLength: 4
          maxLength: 30
      example:
        username: Johndoe
        firstname: John
        lastname: doe
        email: 'johndoe@xyz.com'
    FollowRequest:
      description: 'It which user we want to follow'
      type: object
      properties:
        userID:
          description: "UserID of the person"
          type: integer
          format: int64
      example:
        userID: 123
      required:
        - userID
    BanRequest:
      description: 'It which user we want to ban'
      type: object
      properties:
        userID:
          description: "UserID of the person who we want to ban"
          type: integer
          format: int64
      example:
        userID: 123
      required:
        - userID
    Comment:
      description: 'It contains the text data for the comment on an image'
      type: object
      properties:
        text:
          description: "Text we want to be the comment"
          type: string
          pattern: "^.*?$"
          minLength: 2 
          maxLength: 20
      example:
        text: 'Nice picture'
      required: 
        - text
...
    UserPhotoProfile:
      description: 'It contains details for the photos on a profile of a user'
      type: object
      properties:
        imageID:
          description: "ImageID of the image"
          type: integer
          format: int64
        photo_data:
          description: "Binary photo data"
          type: string
          format: binary
          minLength: 0
          maxLength: 100
        timestamp:
          description: "Timestamp of when the pic is uploaded"
          type: string
          format: date-time
          minLength: 20
          maxLength: 20
      example:
        imageID: 258
        timestamp: "2023-12-21T17:32:28Z"
        photo_data: "010110010"
        #text: 'nice pic'
      required:
        - imageID
        - photo_data
        - timestamp
...
    UserProfile:
      description: 'It contains data for the user profile'
      type: object
      properties:
        userID:
          description: "UserID of the person"
          type: integer
          format: int64
        username:
          description: "Username of the person"
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 15
        firstname:
          description: "Firstname of the person"
          type: string
          pattern: "^.*?$"
          minLength: 3
          maxLength: 15
        lastname:
          description: "Lastname of the person"
          type: string
          pattern: "^.*?$"
          minLength: 3
          maxLength: 15
        email:
          description: "email of the person"
          type: string
          format: email
          minLength: 4
          maxLength: 30
        photos:
          description: "photos on the profile"
          type: array
          minItems: 0
          maxItems: 200
          items:
            $ref: '#/components/schemas/UserPhotoProfile'
      example:
        userID: 123
        username: Johndoe
        firstname: John
        lastname: doe
        email: "johndoe@xyz.com"
        photos:
          imageID: 258
          photo_data: "010110010"
          timestamp: "2023-12-21T17:32:28Z"
        
      required:
        - userID
        - username
        - firstname
        - lastname
        - email
        - photos
  responses:
...
    NotFoundResponseUser:
      description: "User not found"
    NoContentResponse:
      description: "No content found"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /users/{userID}:
    post:
      security:
        - bearerAuth: []
      tags:
        - user
      summary: Update an existing user
      operationId: setMyUserName
      description: 'Update user user can update any one field excpet userID'
      parameters:
        - name: userID
          in: path
          required: true
          description: User ID to be updated
          schema:
            description: 'It sends the userID in path to update user'
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
            example:
                username: Johndoe
                fistname: John
                lastname: doe
                email: 'johndoe@xyz.com'
      responses:
        # '200':
        #   description: User updated successfully
        #   content:
        #     application/json:
        #       schema:
        #         $ref: '#/components/schemas/UserUpdate'
        #       example:
        #         username: John_doe
        #         fistname: John
        #         lastname: doe
        #         email: 'johndoe@xyz.com'
        '201':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUpdate'
              example:
                username: John_doe
                fistname: John
                lastname: doe
                email: 'johndoe@xyz.com'
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'

    delete:
      security:
        - bearerAuth: []
      tags:
        - user
      summary: Delete a user
      operationId: deleteUser
      description: 'Delete a user'
      parameters:
        - name: userID
          in: path
          required: true
          description: User ID to be deleted
          schema:
            description: 'It sends userID from path to delete user'
            type: integer
            format: int64
      responses:
        '204': #chnage this one to 201 and 204 for no contnet 
          description: User deleted successfully
          # content:
          #   application/json:
          #     example:
          #       message: User deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
  /users/{userID}/follow: 
    post:
      security:
        - bearerAuth: []
      summary: Follow a user
      tags:
      - user
      operationId: followUser
      description: 'Follow a user'
      parameters:
      - name: userID
        in: path
        required: true
        description: user id of the person in the session 
        schema:
          description: 'It sends which userID is requestion to follow a user'
          type: integer
          format: int64
      - name: followListOwnerID
        in: query
        required: true
        description: "Person I want to follow"
        schema:
          description: 'It sends userID that I want to follow'
          type: integer
          format: int64

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowRequest'
            example:
                userID: 123

      responses:
        # '201':
        #   description: User followed successfully
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'

    delete:
      security:
        - bearerAuth: []
      summary: Unfollow a user
      tags:
      - user
      operationId: unfollowUser
      description: 'un follow a user'
      parameters:
        - name: userID
          in: path
          required: true
          description: "Follower's user ID"
          schema:
            description: 'It sends userID that is follwoing'
            type: integer
            format: int64
        - name: followListOwnerID
          in: query
          required: true
          description: "User ID of the person to unfollow"
          schema:
            description: 'It sends userID that is being followed'
            type: integer
            format: int64
      responses:
        #no response for 200 
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'
  /users/{userID}/ban:
    post:
      security:
        - bearerAuth: []
      summary: Ban a user
      tags:
      - user
      operationId:  banUser
      description: 'Ban a user'
      parameters:
        - name: userID
          in: path
          required: true
          description: "User ID of the person in the seeion right now"
          schema:
            description: 'User who is in the session'
            type: integer
            format: int64
        - name: BannedUserID
          in: query
          required: true
          description: "User ID to be banned"
          schema:
            description: 'User ID to ban'
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BanRequest'
            example:
                userID: 123

      responses:
        # '200':
        #   description: User banned successfully
        # '201':
        #   description: User banned successfully
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'

    delete:
      security:
        - bearerAuth: []
      summary: Unban a user
      tags:
      - user
      operationId: unbanUser
      description: 'un ban a banned user'
      parameters:
        - name: userID
          in: path
          required: true
          description: "User ID to be unbanned"
          schema:
            description: 'User ID who has been banned'
            type: integer
            format: int64
            
        - name: BannedUserID #is the UserID of the person I want to ban
          in: query
          required: true
          description: "User ID of the person to ban"
          schema:
            description: 'User ID of the person in the session'
            type: integer
            format: int64

          
      responses:
        # '200':
        #   description: User unbanned successfully
        
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'
...
  /users/{userID}/image/{imageID}:
    delete:
      security:
        - bearerAuth: []
      summary: Delete an image
      tags:
      - image
      operationId: deletePhoto
      description: 'Delete a picture'
      parameters:
        - name: userID
          in: path
          required: true
          description: User ID deleting the image
          schema:
            description: "User ID of the person in the session deleting his own photo"
            type: integer
            format: int64
        - name: imageID
          in: path
          required: true
          description: Image ID to be deleted
          schema:
            description: "Image ID of the picture to be deleted"
            type: integer
            format: int64
      responses:
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'
...
  /users/{userID}/image/{imageID}/comment/stream:
    get:
      security:
        - bearerAuth: []
      summary: Get all comments on an image
      tags:
      - frontPage
      operationId: getAllComments
      description: 'Get all comments on a image'
      parameters:
        - name: userID
          in: path
          required: true
          description: "User ID whose profile is being requested"
          schema:
            description: "User ID of the person who we are trying to follow"
            type: integer
            format: int64
        - name: imageID
          in: path
          required: true
          description: "Image ID whose profile is being requested"
          schema:
            description: "Image ID aka the picture to which all the comments are linked to"
            type: integer
            format: int64
      responses:
        '200':
          description: Comment on image retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
...
  /users/{userID}/profile:
    get:
      security:
        - bearerAuth: []
      summary: Get user profile
      tags:
      - frontPage
      operationId: getUserProfile
      description: ' Get the user profile '
      parameters:
        - name: userID
          in: path
          required: true
          description: "User ID whose profile is being requested"
          schema:
            description: "User ID of the person for which we want to see the profile"
            type: integer
            format: int64
      responses:
        '200':
          description: "User profile retrieved successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              example:
                userID: 123
                username: jondoe
                firstname: Jon
                lastname: doe
                email: "johndoe@xyz.com"
                photos:
                  - imageID: 1
                    photo_data: "010110010"
                    timestamp: "2023-01-01T12:00:00Z"
                  - imageID: 2
                    photo_data: "010110011"
                    timestamp: "2023-01-02T12:30:00Z"
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
...